Music and Timer
Music plays automatically when start up game. It's just an audio source attached to Sound Object, but you can put an audio source anywhere.
The timer is a an object called BarTimer Test Screen in the assets folder. It's not in the scene bc it was in the way. Drag it into the scene. Press T to toggle.