Unit Store in Unit Bench Attempt
UnitStore.cs in the Script folder is attached to the UnitStore GameObject (an empty object), which is nested in the Grid. The script instantiates (summons) a random Unit object (one of the animated guys) on the Unit Bench for the length of the unit bench (using the new GetWidth() function in UnitBench). The position is determined by the new GetCellCenter() function in UnitBenchMeshGenerator()